+++ services/radio-manager-backend/src/models/types.rs
@@ -1,32 +1,36 @@
+use serde::Serialize;
 use std::ops::Deref;
 
-#[derive(Serialize, Clone)]
-pub(crate) struct UserId(usize);
+#[derive(Serialize, Clone, sqlx::Type)]
+#[sqlx(transparent)]
+pub(crate) struct UserId(i32);
 
 impl Deref for UserId {
-    type Target = usize;
+    type Target = i32;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-#[derive(Serialize, Clone)]
-pub(crate) struct TrackId(usize);
+#[derive(Serialize, Clone, sqlx::Type)]
+#[sqlx(transparent)]
+pub(crate) struct TrackId(i32);
 
 impl Deref for TrackId {
-    type Target = usize;
+    type Target = i32;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-#[derive(Serialize, Clone)]
-pub(crate) struct FileId(usize);
+#[derive(Serialize, Clone, sqlx::Type)]
+#[sqlx(transparent)]
+pub(crate) struct FileId(i32);
 
 impl Deref for FileId {
-    type Target = usize;
+    type Target = i32;
 
     fn deref(&self) -> &Self::Target {
         &self.0
